Move AST related entities into the alogic.ast package
